feat: implement live_saturation calculation
- Added  dependency for improved numerical precision in financial calculations.
- Updated  to use  for , enhancing accuracy.
- Introduced  struct to encapsulate protocol parameters in epoch state queries.
- Enhanced parameters state handling to retrieve and respond with the latest epoch parameters.
- Refactored REST handlers to incorporate new epoch state queries, improving data retrieval for pools.

+++ modules/rest_blockfrost/src/types.rs
@@ -1,4 +1,5 @@
 use acropolis_common::VotingProcedure;
+use rust_decimal::Decimal;
 use serde::Serialize;
 
 /// REST response structure for proposal votes
@@ -12,10 +13,10 @@ pub struct VoteRest {
 pub struct PoolExtendedRest {
     pub pool_id: String,
     pub hex: String,
-    pub active_stake: String,  // u64 in string
-    pub live_stake: String, // u64 in string
+    pub active_stake: String, // u64 in string
+    pub live_stake: String,   // u64 in string
     pub blocks_minted: usize,
-    pub live_saturation: f32,
+    pub live_saturation: Decimal,
     pub declared_pledge: String, // u64 in string
     pub margin_cost: f32,
     pub fixed_cost: String, // u64 in string
